* Fixed Merging Method Into One In Repositories
